requirements.txt
clipboard_monitor==1.0.5
keyboard==0.13.5
Pillow==11.0.0
pyperclip==1.9.0
pytesseract==0.3.13
